pyproject.toml
[project]
name = "mcp-crypto-portfolio"
version = "0.1.0"
description = "MCP server for crypto portfolio using KuCoin + Notion"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastmcp>=2.0.0",
  "httpx>=0.27",
  "pydantic>=2.8",
  "python-dotenv>=1.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

# If your code lives under src/
[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
